Code Cleanup + Styling Fixes + Image Cache (#417)
* Fix erroneous error log

* swiftlint :/

* Styling improvements + Removed NSHomeDirectory()

* More cleanup

* URL Refactor

* More cleanup

* Installer refactor

* Better error checking

* Image Cache

* Fix AppSettings

* Update IPA Library

* Clear Cache in Menu Bar

* Fix resizable

* Remove extra VStack

* Fix icons not refreshing

* Fixed Infinity Blade Icon

* Fixed Duolingo Icon

* Stops unnecessary copying of icons in bundle

* Use Proper Cache Directory

+++ PlayCover/Views/App Views/PlayAppView.swift
@@ -167,9 +167,11 @@ struct PlayAppView: View {
                 .appendingPathComponent("Cookies")
                 .appendingPathComponent("Cookies")
                 .appendingPathExtension("binarycookies")
-            try FileManager.default.removeItem(at: cookieURL)
+            if FileManager.default.fileExists(atPath: cookieURL.path) {
+                try FileManager.default.removeItem(at: cookieURL)
+            }
         } catch {
-            Log.shared.error(error)
+            print("Error when attempting to remove Twitter session cookie: \(error)")
         }
     }
 }
@@ -179,56 +181,58 @@ struct PlayAppConditionalView: View {
     @Binding var selectedTextColor: Color
     @Binding var selected: PlayApp?
 
+    @State var iconURL: URL?
     @State var app: PlayApp
     @State var isList: Bool
 
     var body: some View {
-        if isList {
-            HStack(alignment: .center, spacing: 0) {
-                if let img = app.icon {
-                    Image(nsImage: img)
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .frame(width: 30, height: 30)
-                        .cornerRadius(7.5)
-                        .shadow(radius: 1)
-                        .padding(.horizontal, 15)
-                        .padding(.vertical, 5)
-                } else {
-                    ProgressView()
-                        .progressViewStyle(.circular)
-                        .frame(width: 60, height: 60)
-                }
-                Text(app.name)
-                    .foregroundColor(selected?.info.bundleIdentifier == app.info.bundleIdentifier ?
-                                     selectedTextColor : Color.primary)
-                Spacer()
-                Text(app.settings.info.bundleVersion)
-                    .padding(.horizontal, 15)
-                    .foregroundColor(.secondary)
-            }
-            .contentShape(Rectangle())
-            .background(
-                RoundedRectangle(cornerRadius: 4)
-                    .fill(selected?.info.bundleIdentifier == app.info.bundleIdentifier ?
-                        selectedBackgroundColor : Color.clear)
-                    .brightness(-0.2)
-                )
-        } else {
-            VStack(alignment: .center, spacing: 0) {
-                VStack {
-                    if let img = app.icon {
-                        Image(nsImage: img)
+        Group {
+            if isList {
+                HStack(alignment: .center, spacing: 0) {
+                    AsyncImage(url: iconURL) { image in
+                        image
                             .resizable()
                             .aspectRatio(contentMode: .fit)
-                            .frame(width: 60, height: 60)
-                            .cornerRadius(15)
+                            .frame(width: 30, height: 30)
+                            .cornerRadius(7.5)
                             .shadow(radius: 1)
-                    } else {
+                            .padding(.horizontal, 15)
+                            .padding(.vertical, 5)
+                    } placeholder: {
                         ProgressView()
                             .progressViewStyle(.circular)
                             .frame(width: 60, height: 60)
                     }
+
+                    Text(app.name)
+                        .foregroundColor(selected?.info.bundleIdentifier == app.info.bundleIdentifier ?
+                                         selectedTextColor : Color.primary)
+                    Spacer()
+                    Text(app.settings.info.bundleVersion)
+                        .padding(.horizontal, 15)
+                        .foregroundColor(.secondary)
+                }
+                .contentShape(Rectangle())
+                .background(
+                    RoundedRectangle(cornerRadius: 4)
+                        .fill(selected?.info.bundleIdentifier == app.info.bundleIdentifier ?
+                            selectedBackgroundColor : Color.clear)
+                        .brightness(-0.2)
+                    )
+            } else {
+                VStack {
+                    AsyncImage(url: iconURL) { image in
+                        image
+                            .resizable()
+                            .aspectRatio(contentMode: .fit)
+                            .cornerRadius(15)
+                            .shadow(radius: 1)
+                    } placeholder: {
+                        ProgressView()
+                            .progressViewStyle(.circular)
+                    }
+                    .frame(width: 60, height: 60)
+
                     Text(app.name)
                         .lineLimit(1)
                         .multilineTextAlignment(.center)
@@ -244,8 +248,13 @@ struct PlayAppConditionalView: View {
                             )
                         .frame(width: 130, height: 20)
                 }
+                .frame(width: 130, height: 130)
             }
-            .frame(width: 130, height: 130)
+        }
+        .task(priority: .userInitiated) {
+            iconURL = ImageCache.getLocalImageURL(bundleID: app.info.bundleIdentifier,
+                                                  bundleURL: app.url,
+                                                  primaryIconName: app.info.primaryIconName)
         }
     }
 }
